kd: rename lazy -> kloud
